Ensure that ports can be disabled
The previous approach, where an _empty_ value for the configuration property that controls on what TCP port the webserver will be listening, caused a NullPointerException to be thrown.

In this commit, the 'disbled' state is represented by the value `-1` (rather than a missing value) in of the corresponding system properties. In the frontend, the value is still shown as absent, as that's more likely to be understood correctly by end-users.

fixes #1